Refactor of matrices
